# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/auth/register": {
      "post": {
        "operationId": "register",
        "parameters": [
          {
            "name": "user-agent",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthStatusDto"
                }
              }
            }
          }
        },
        "summary": "Cria uma conta de cliente e inicia a sessão",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "operationId": "login",
        "parameters": [
          {
            "name": "user-agent",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthStatusDto"
                }
              }
            }
          }
        },
        "summary": "Login com e-mail e senha (cookie de sessão httpOnly)",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/v1/auth/forgot-password": {
      "post": {
        "operationId": "forgotPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForgotPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkDto"
                }
              }
            }
          }
        },
        "summary": "Envia link de redefinição de senha (resposta idêntica mesmo se o e-mail não existir)",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/reset-password": {
      "post": {
        "operationId": "resetPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkDto"
                }
              }
            }
          }
        },
        "summary": "Define nova senha a partir do token recebido por e-mail",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/change-password": {
      "post": {
        "operationId": "changePassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Troca a senha do usuário logado (encerra as outras sessões)",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/mfa/setup": {
      "post": {
        "operationId": "mfaSetup",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MfaSetupResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Gera segredo TOTP e QR code para ativar o segundo fator",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/mfa/enable": {
      "post": {
        "operationId": "mfaEnable",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MfaCodeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthStatusDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Confirma o código do app autenticador e ativa o MFA",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/mfa/verify": {
      "post": {
        "operationId": "mfaVerify",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MfaCodeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthStatusDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Verifica o código TOTP da sessão atual",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/profile": {
      "patch": {
        "operationId": "updateProfile",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfileDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Atualiza nome, telefone, CPF e preferência de newsletter",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/v1/checkout/quote": {
      "post": {
        "operationId": "quote",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CartQuoteInputDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CartQuoteDto"
                }
              }
            }
          }
        },
        "summary": "Recalcula a sacola no servidor: preços atuais, estoque, cupom, frete e parcelas",
        "tags": [
          "checkout"
        ]
      }
    },
...
    "/api/v1/checkout/orders": {
      "post": {
        "operationId": "placeOrder",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckoutInputDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckoutResultDto"
                }
              }
            }
          }
        },
        "summary": "Cria o pedido, reserva estoque e gera a cobrança (Pix, boleto ou cartão)",
        "tags": [
          "checkout"
        ]
      }
    },
...
  },
  "info": {
    "title": "Vellor API",
    "description": "API da loja Vellor (relógios de luxo e perfumes de nicho). Autenticação por cookie de sessão httpOnly; erros no formato RFC 7807 (application/problem+json).",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "vellor_sid"
      }
    },
    "schemas": {
...
      "RegisterDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 3,
            "maxLength": 120
          },
          "email": {
            "type": "string",
            "maxLength": 160,
            "format": "email",
            "pattern": "^(?!\\.)(?!.*\\.\\.)([A-Za-z0-9_'+\\-\\.]*)[A-Za-z0-9_+-]@([A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$"
          },
          "password": {
            "type": "string",
            "minLength": 10,
            "maxLength": 128
          },
          "phone": {
            "type": "string"
          },
          "acceptTerms": {
            "type": "boolean",
            "enum": [
              true
            ]
          },
          "newsletterOptIn": {
            "default": false,
            "type": "boolean"
          }
        },
        "required": [
          "name",
          "email",
          "password",
          "acceptTerms"
        ]
      },
      "AuthStatusDto": {
        "type": "object",
        "properties": {
          "authenticated": {
            "type": "boolean"
          },
          "user": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "format": "uuid",
                "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
              },
              "name": {
                "type": "string"
              },
              "email": {
                "type": "string"
              },
              "role": {
                "type": "string",
                "enum": [
                  "customer",
                  "admin"
                ]
              },
              "phone": {
                "type": [
                  "string",
                  "null"
                ]
              },
              "cpfMasked": {
                "type": [
                  "string",
                  "null"
                ]
              },
              "mfaEnabled": {
                "type": "boolean"
              },
              "mfaVerified": {
                "type": "boolean"
              },
              "newsletterOptIn": {
                "type": "boolean"
              },
              "createdAt": {
                "type": "string"
              }
            },
            "required": [
              "id",
              "name",
              "email",
              "role",
              "phone",
              "cpfMasked",
              "mfaEnabled",
              "mfaVerified",
              "newsletterOptIn",
              "createdAt"
            ],
            "nullable": true
          },
          "mfaSetupRequired": {
            "type": "boolean"
          },
          "mfaPending": {
            "type": "boolean"
          }
        },
        "required": [
          "authenticated",
          "user",
          "mfaSetupRequired",
          "mfaPending"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "maxLength": 160,
            "format": "email",
            "pattern": "^(?!\\.)(?!.*\\.\\.)([A-Za-z0-9_'+\\-\\.]*)[A-Za-z0-9_+-]@([A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$"
          },
          "password": {
            "type": "string",
            "minLength": 1,
            "maxLength": 128
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "OkDto": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean",
            "enum": [
              true
            ]
          }
        },
        "required": [
          "ok"
        ]
      },
      "ForgotPasswordDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "maxLength": 160,
            "format": "email",
            "pattern": "^(?!\\.)(?!.*\\.\\.)([A-Za-z0-9_'+\\-\\.]*)[A-Za-z0-9_+-]@([A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$"
          }
        },
        "required": [
          "email"
        ]
      },
      "ResetPasswordDto": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "minLength": 20,
            "maxLength": 200
          },
          "password": {
            "type": "string",
            "minLength": 10,
            "maxLength": 128
          }
        },
        "required": [
          "token",
          "password"
        ]
      },
      "ChangePasswordDto": {
        "type": "object",
        "properties": {
          "currentPassword": {
            "type": "string",
            "minLength": 1,
            "maxLength": 128
          },
          "newPassword": {
            "type": "string",
            "minLength": 10,
            "maxLength": 128
          }
        },
        "required": [
          "currentPassword",
          "newPassword"
        ]
      },
      "MfaSetupResponseDto": {
        "type": "object",
        "properties": {
          "otpauthUrl": {
            "type": "string"
          },
          "qrCodeDataUrl": {
            "type": "string"
          },
          "secretMasked": {
            "type": "string"
          }
        },
        "required": [
          "otpauthUrl",
          "qrCodeDataUrl",
          "secretMasked"
        ]
      },
      "MfaCodeDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          }
        },
        "required": [
          "code"
        ]
      },
      "UpdateProfileDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 3,
            "maxLength": 120
          },
          "phone": {
            "type": "string"
          },
          "cpf": {
            "type": "string"
          },
          "newsletterOptIn": {
            "type": "boolean"
          }
        }
      },
      "MeDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "customer",
              "admin"
            ]
          },
          "phone": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "cpfMasked": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "mfaEnabled": {
            "type": "boolean"
          },
          "mfaVerified": {
            "type": "boolean"
          },
          "newsletterOptIn": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "email",
          "role",
          "phone",
          "cpfMasked",
          "mfaEnabled",
          "mfaVerified",
          "newsletterOptIn",
          "createdAt"
        ]
      },
...
      "CartQuoteInputDto": {
        "type": "object",
        "properties": {
          "items": {
            "minItems": 1,
            "maxItems": 20,
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "variantId": {
                  "type": "string",
                  "format": "uuid",
                  "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                },
                "quantity": {
                  "type": "integer",
                  "minimum": 1,
                  "maximum": 10
                }
              },
              "required": [
                "variantId",
                "quantity"
              ]
            }
          },
          "couponCode": {
            "type": "string",
            "maxLength": 32
          },
          "cep": {
            "type": "string"
          },
          "shippingService": {
            "type": "string",
            "enum": [
              "PAC",
              "SEDEX"
            ]
          }
        },
        "required": [
          "items"
        ]
      },
...
      "CheckoutInputDto": {
        "type": "object",
        "properties": {
          "items": {
            "minItems": 1,
            "maxItems": 20,
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "variantId": {
                  "type": "string",
                  "format": "uuid",
                  "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                },
                "quantity": {
                  "type": "integer",
                  "minimum": 1,
                  "maximum": 10
                }
              },
              "required": [
                "variantId",
                "quantity"
              ]
            }
          },
          "customer": {
            "type": "object",
            "properties": {
              "name": {
                "type": "string",
                "minLength": 3,
                "maxLength": 120
              },
              "email": {
                "type": "string",
                "maxLength": 160,
                "format": "email",
                "pattern": "^(?!\\.)(?!.*\\.\\.)([A-Za-z0-9_'+\\-\\.]*)[A-Za-z0-9_+-]@([A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$"
              },
              "phone": {
                "type": "string"
              },
              "cpf": {
                "type": "string"
              }
            },
            "required": [
              "name",
              "email",
              "phone",
              "cpf"
            ]
          },
          "shippingAddress": {
            "type": "object",
            "properties": {
              "recipientName": {
                "type": "string",
                "minLength": 3,
                "maxLength": 120
              },
              "cep": {
                "type": "string"
              },
              "street": {
                "type": "string",
                "minLength": 2,
                "maxLength": 160
              },
              "number": {
                "type": "string",
                "minLength": 1,
                "maxLength": 20
              },
              "complement": {
                "type": "string",
                "maxLength": 80,
                "nullable": true
              },
              "district": {
                "type": "string",
                "minLength": 2,
                "maxLength": 80
              },
              "city": {
                "type": "string",
                "minLength": 2,
                "maxLength": 80
              },
              "state": {
                "type": "string",
                "enum": [
                  "AC",
                  "AL",
                  "AP",
                  "AM",
                  "BA",
                  "CE",
                  "DF",
                  "ES",
                  "GO",
                  "MA",
                  "MT",
                  "MS",
                  "MG",
                  "PA",
                  "PB",
                  "PR",
                  "PE",
                  "PI",
                  "RJ",
                  "RN",
                  "RS",
                  "RO",
                  "RR",
                  "SC",
                  "SP",
                  "SE",
                  "TO"
                ]
              },
              "reference": {
                "type": "string",
                "maxLength": 160,
                "nullable": true
              }
            },
            "required": [
              "cep",
              "street",
              "number",
              "district",
              "city",
              "state"
            ]
          },
          "shippingService": {
            "type": "string",
            "enum": [
              "PAC",
              "SEDEX"
            ]
          },
          "payment": {
            "type": "object",
            "properties": {
              "method": {
                "type": "string",
                "enum": [
                  "pix",
                  "boleto",
                  "credit_card"
                ]
              },
              "installments": {
                "default": 1,
                "type": "integer",
                "minimum": 1,
                "maximum": 12
              }
            },
            "required": [
              "method"
            ]
          },
          "couponCode": {
            "type": "string",
            "maxLength": 32
          },
          "notes": {
            "type": "string",
            "maxLength": 500
          },
          "acceptTerms": {
            "type": "boolean",
            "enum": [
              true
            ]
          },
          "saveAddress": {
            "type": "boolean"
          }
        },
        "required": [
          "items",
          "customer",
          "shippingAddress",
          "shippingService",
          "payment",
          "acceptTerms"
        ]
      },
      "CheckoutResultDto": {
        "type": "object",
        "properties": {
          "orderId": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          },
          "orderNumber": {
            "type": "string"
          },
          "accessToken": {
            "type": "string"
          },
          "payment": {
            "type": "object",
            "properties": {
              "method": {
                "type": "string",
                "enum": [
                  "pix",
                  "boleto",
                  "credit_card"
                ]
              },
              "status": {
                "type": "string",
                "enum": [
                  "pending",
                  "confirmed",
                  "received",
                  "overdue",
                  "refunded",
                  "cancelled"
                ]
              },
              "installments": {
                "type": "integer",
                "minimum": -9007199254740991,
                "maximum": 9007199254740991
              },
              "installmentCents": {
                "type": "integer",
                "minimum": -9007199254740991,
                "maximum": 9007199254740991
              },
              "pix": {
                "type": "object",
                "properties": {
                  "payload": {
                    "type": "string"
                  },
                  "qrCodeBase64": {
                    "type": [
                      "string",
                      "null"
                    ]
                  },
                  "expiresAt": {
                    "type": [
                      "string",
                      "null"
                    ]
                  }
                },
                "required": [
                  "payload",
                  "qrCodeBase64",
                  "expiresAt"
                ],
                "nullable": true
              },
              "boleto": {
                "type": "object",
                "properties": {
                  "url": {
                    "type": "string"
                  },
                  "identificationField": {
                    "type": [
                      "string",
                      "null"
                    ]
                  },
                  "dueDate": {
                    "type": "string"
                  }
                },
                "required": [
                  "url",
                  "identificationField",
                  "dueDate"
                ],
                "nullable": true
              },
              "creditCard": {
                "type": "object",
                "properties": {
                  "checkoutUrl": {
                    "type": "string"
                  }
                },
                "required": [
                  "checkoutUrl"
                ],
                "nullable": true
              },
              "paidAt": {
                "type": [
                  "string",
                  "null"
                ]
              }
            },
            "required": [
              "method",
              "status",
              "installments",
              "installmentCents",
              "pix",
              "boleto",
              "creditCard",
              "paidAt"
            ]
          }
        },
        "required": [
          "orderId",
          "orderNumber",
          "accessToken",
          "payment"
        ]
      },
...
    }
  }
}